Fix up polynomial error (missing loop over degree) and integrate it throughout
