+++ include/lagrange/lagrange.h
@@ -6,6 +6,7 @@
 #include <array>
 #include <cmath>
 #include <vector>
+#include <math_utils/polynomials.h>
 
 namespace local_lagrange {
 
@@ -24,9 +25,9 @@ arma::mat computeInterpolationMatrix(const arma::mat& centers,
       interp_matrix(row, col) = interp_matrix(col, row) = kernel(dist);
     }
   }
-
-  mathtools::buildPolynomialMatrix<Dimension, 1>(
-      interp_matrix, centers);  ///@todo srowe: Degree should be computable
+  
+  ///@todo srowe: Degree should be computable
+  buildPolynomialMatrix<Dimension, 1>(interp_matrix, centers); 
 
   return interp_matrix;
 }
